Add direct switching to multiple alternates
Added a focus option to the request for this extension.  That focus option matches a nested mappings object in the standard format.  Each nested mappings object has a single match resulting in an immediate switch.  The removes the need for the intermediary chooser most of the time.  The default mappings are at mappings.all and can still be helpful for splits/tabs/peeks etc.